clear command with role, bot following the first user who enters in voice channel, bot following the last user who changes to other voice channel, bot leaving the voice channel when the last user leaves

+++ src/main/java/me/grandeaposta/discordbot/command/commands/ClearCommand.java
@@ -2,6 +2,7 @@ package me.grandeaposta.discordbot.command.commands;
 
 import java.util.List;
 
+import me.grandeaposta.discordbot.Config;
 import me.grandeaposta.discordbot.command.CommandContext;
 import me.grandeaposta.discordbot.command.ICommand;
 import net.dv8tion.jda.api.entities.Message;
@@ -16,18 +17,20 @@ public class ClearCommand implements ICommand {
 
     @Override
     public void handle(CommandContext ctx) {
-        MessageReceivedEvent event = ctx.getEvent();
-
         
-
-        List<String> args = ctx.getArgs();
-
-        if (!isNumber(args.get(0))) {
-            return;
+        if(ctx.getEvent().getMember().getRoles().toString().contains(Config.get("role_principal"))){
+            MessageReceivedEvent event = ctx.getEvent();
+    
+            List<String> args = ctx.getArgs();
+    
+            if (!isNumber(args.get(0))) {
+                return;
+            }
+    
+            List<Message> messageList = event.getChannel().getHistory().retrievePast(Integer.parseInt(args.get(0)) + 1).complete();
+            event.getGuildChannel().deleteMessages(messageList).queue();            
         }
 
-        List<Message> messageList = event.getChannel().getHistory().retrievePast(Integer.parseInt(args.get(0)) + 1).complete();
-        event.getGuildChannel().deleteMessages(messageList).queue();
     }
 
     public boolean isNumber(String msg) {
